// Admin/src/app/services/models/campus.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TrainingCenter } from '../models/training-center';
export interface Campus {
  actived?: boolean;
  archived?: boolean;
  capacity?: number;
  createdBy?: number;
  createdDate?: string;
  id?: number;
  lastModifiedBy?: number;
  lastModifiedDate?: string;
  name?: string;
  quarter?: string;
  town?: string;
  trainingCenter?: TrainingCenter;
  xcoor?: number;
  ycoor?: number;
}